chore: adjust the method of obtaining the mcp prompt
