+++ projects/electron-app/src/main/ipc/mcpHandlers.ts
@@ -1,6 +1,6 @@
 import { ipcMain } from 'electron';
 import { IPC_EVENTS, ResultUtil } from '@evo/utils';
-import { BaseResult, IMcpCategoryMeta, IMcpMeta } from '@evo/types';
+import { BaseResult, IGetMcpToolParams, IMcpCategoryMeta, IMcpMeta, IMcpService } from '@evo/types';
 import { managerService } from '../services/ManagerService';
 
 export async function setupMcpHandlers() {
@@ -117,14 +117,10 @@ export async function setupMcpHandlers() {
     }
   );
 
-  ipcMain.handle(IPC_EVENTS.MCP.GET_TOOLS, async (_, mcpId: string): Promise<BaseResult<any[]>> => {
-    return MCPClientManager.getTools(mcpId);
-  });
-
   ipcMain.handle(
-    IPC_EVENTS.MCP.GET_MCP_PROMPT,
-    async (_, mcpIds: string[], userPrompt: string): Promise<BaseResult<string>> => {
-      return MCPClientManager.getMcpPrompt(mcpIds, userPrompt);
+    IPC_EVENTS.MCP.GET_TOOLS,
+    async (_, params: IGetMcpToolParams): Promise<BaseResult<any[]>> => {
+      return MCPClientManager.getTools(params);
     }
   );
